[CHORE/#202] 패키지명 변경
