+++ app/src/main/java/com/spoony/spoony/presentation/gourmet/map/component/MapTopAppBar.kt
@@ -1,4 +1,4 @@
-package com.spoony.spoony.presentation.map.map.component
+package com.spoony.spoony.presentation.gourmet.map.component
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
