+++ app/src/main/java/com/spoony/spoony/presentation/gourmet/map/navigaion/MapNavigation.kt
@@ -1,4 +1,4 @@
-package com.spoony.spoony.presentation.map.map.navigaion
+package com.spoony.spoony.presentation.gourmet.map.navigaion
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.spoony.spoony.core.navigation.MainTabRoute
-import com.spoony.spoony.presentation.map.map.MapRoute
+import com.spoony.spoony.presentation.gourmet.map.MapRoute
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateToMap(
