+++ app/src/main/java/com/spoony/spoony/presentation/gourmet/search/MapSearchViewModel.kt
@@ -1,10 +1,10 @@
-package com.spoony.spoony.presentation.map.search
+package com.spoony.spoony.presentation.gourmet.search
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.spoony.spoony.core.state.UiState
 import com.spoony.spoony.domain.repository.MapRepository
-import com.spoony.spoony.presentation.map.map.model.toModel
+import com.spoony.spoony.presentation.gourmet.map.model.toModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 import kotlinx.collections.immutable.persistentListOf
