+++ app/src/main/java/com/spoony/spoony/presentation/gourmet/search/component/MapSearchRecentEmptyScreen.kt
@@ -1,7 +1,8 @@
-package com.spoony.spoony.presentation.map.map.component.bottomsheet
+package com.spoony.spoony.presentation.gourmet.search.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
@@ -9,42 +10,34 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.spoony.spoony.R
-import com.spoony.spoony.core.designsystem.component.button.SpoonyButton
 import com.spoony.spoony.core.designsystem.theme.SpoonyAndroidTheme
-import com.spoony.spoony.core.designsystem.type.ButtonSize
-import com.spoony.spoony.core.designsystem.type.ButtonStyle
 
 @Composable
-fun MapEmptyBottomSheetContent(
-    onClick: () -> Unit,
+fun MapSearchRecentEmptyScreen(
     modifier: Modifier = Modifier
 ) {
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
         modifier = modifier
-            .padding(vertical = 24.dp)
             .fillMaxWidth()
+            .padding(horizontal = 20.dp, vertical = 72.dp)
     ) {
         Image(
-            painter = painterResource(R.drawable.img_empty_home),
-            contentDescription = null
+            painter = painterResource(R.drawable.img_map_search_recent_empty),
+            contentDescription = null,
+            modifier = Modifier
+                .fillMaxWidth()
+                .aspectRatio(220 / 100f)
+                .padding(horizontal = 50.dp)
         )
         Text(
-            text = "아직 추가된 장소가 없어요.\n다른 사람의 리스트를 떠먹어보세요!",
+            text = "구체적인 장소를 검색해 보세요",
             style = SpoonyAndroidTheme.typography.body2m,
             color = SpoonyAndroidTheme.colors.gray500,
-            textAlign = TextAlign.Center,
             modifier = Modifier
-                .padding(vertical = 16.dp)
-        )
-        SpoonyButton(
-            text = "떠먹으러 가기",
-            size = ButtonSize.Xsmall,
-            style = ButtonStyle.Secondary,
-            onClick = onClick
+                .padding(top = 24.dp)
         )
     }
 }
