+++ app/src/main/java/com/spoony/spoony/presentation/gourmet/search/navigation/MapSearchNavigation.kt
@@ -1,4 +1,4 @@
-package com.spoony.spoony.presentation.map.search.navigation
+package com.spoony.spoony.presentation.gourmet.search.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.spoony.spoony.core.navigation.Route
-import com.spoony.spoony.presentation.map.search.MapSearchRoute
+import com.spoony.spoony.presentation.gourmet.search.MapSearchRoute
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateToMapSearch(
