+++ app/src/main/java/com/spoony/spoony/presentation/main/MainNavigator.kt
@@ -10,9 +10,9 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navOptions
 import com.spoony.spoony.presentation.explore.navigation.navigateToExplore
-import com.spoony.spoony.presentation.map.map.navigaion.Map
-import com.spoony.spoony.presentation.map.map.navigaion.navigateToMap
-import com.spoony.spoony.presentation.map.search.navigation.navigateToMapSearch
+import com.spoony.spoony.presentation.gourmet.map.navigaion.Map
+import com.spoony.spoony.presentation.gourmet.map.navigaion.navigateToMap
+import com.spoony.spoony.presentation.gourmet.search.navigation.navigateToMapSearch
 import com.spoony.spoony.presentation.placeDetail.navigation.navigateToPlaceDetail
 import com.spoony.spoony.presentation.register.navigation.navigateToRegister
 import com.spoony.spoony.presentation.report.navigation.navigateToReport
