+++ app/src/main/java/com/spoony/spoony/presentation/main/MainScreen.kt
@@ -20,9 +20,9 @@ import androidx.navigation.compose.NavHost
 import com.spoony.spoony.core.designsystem.component.snackbar.TextSnackbar
 import com.spoony.spoony.core.designsystem.event.LocalSnackBarTrigger
 import com.spoony.spoony.presentation.explore.navigation.exploreNavGraph
+import com.spoony.spoony.presentation.gourmet.map.navigaion.mapNavGraph
+import com.spoony.spoony.presentation.gourmet.search.navigation.mapSearchNavGraph
 import com.spoony.spoony.presentation.main.component.MainBottomBar
-import com.spoony.spoony.presentation.map.map.navigaion.mapNavGraph
-import com.spoony.spoony.presentation.map.search.navigation.mapSearchNavGraph
 import com.spoony.spoony.presentation.placeDetail.navigation.placeDetailNavGraph
 import com.spoony.spoony.presentation.register.navigation.registerNavGraph
 import com.spoony.spoony.presentation.report.navigation.reportNavGraph
