+++ app/src/main/java/com/spoony/spoony/presentation/main/MainTab.kt
@@ -6,7 +6,7 @@ import com.spoony.spoony.R
 import com.spoony.spoony.core.navigation.MainTabRoute
 import com.spoony.spoony.core.navigation.Route
 import com.spoony.spoony.presentation.explore.navigation.Explore
-import com.spoony.spoony.presentation.map.map.navigaion.Map
+import com.spoony.spoony.presentation.gourmet.map.navigaion.Map
 import com.spoony.spoony.presentation.register.navigation.Register
 
 enum class MainTab(
